fbsdk fixes
