+++ app/(public)/pre-signin.web.tsx
@@ -18,10 +18,8 @@ import {
   DUMMY_USER_CREDENTIALS2,
 } from "@/constants/User";
 import Colors from "@/constants/Colors";
-// import { LoginManager } from "react-native-fbsdk-next";
 import { FirestoreDB } from "@/utils/firestore";
 import { collection, doc, getDoc, setDoc } from "firebase/firestore";
-// import { AccessToken } from "react-native-fbsdk-next";
 
 WebBrowser.maybeCompleteAuthSession();
 
@@ -97,27 +95,7 @@ const PreSignIn = () => {
   };
 
   const handleFacebookSignIn = async () => {
-    if (Platform.OS === "web") {
-      await promptAsync();
-    }
-    // } else if (Platform.OS === "android" || Platform.OS === "ios") {
-    //   LoginManager.logInWithPermissions(["public_profile"]).then(
-    //     function (result) {
-    //       if (result.isCancelled) {
-    //       } else {
-    //         AccessToken.getCurrentAccessToken().then((data) => {
-    //           const token = data?.accessToken;
-    //           if (token) {
-    //             handleFirebaseSignIn(token);
-    //           }
-    //         });
-    //       }
-    //     },
-    //     function (error) {
-    //       console.log("==> Login fail with error: " + error);
-    //     }
-    //   );
-    // }
+    await promptAsync();
   };
 
   const renderSocialButton = (
@@ -158,12 +136,12 @@ const PreSignIn = () => {
             <Paragraph style={styles.paragraph}>{slide.text}</Paragraph>
             {slide.key === "connect" && (
               <View style={styles.socialContainer}>
-                {/* {renderSocialButton(
+                {renderSocialButton(
                   "logo-facebook",
                   "Login with Facebook",
                   // () => promptAsync({}),
                   () => handleFacebookSignIn()
-                )} */}
+                )}
                 {renderSocialButton(
                   "mail-outline",
                   "Login with Email",
